requirements.txt
streamlit
scikit-learn
Pillow
numpy
